requirements.txt
aiogram==3.6.0
fastapi==0.115.5
uvicorn==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.36
asyncpg==0.29.0
pydantic==2.9.2
pytz==2024.1
